Add memory read based on paddr and removed revert
After translating the virtual address and finding the pma_entry we're
finally equiped to read memory for the instruction (?) value. The code
reads the value on memory manager at address paddr. Have to make sure
that there is indeed the insn value there.
Also removed the rever() if vaddr == 0. The emulator is never supposed
to revert - even if the machine is not built correctly.