Reverting switch from uint64 to bytes8 on MM
To make development easier we had changed the values on mmInstantiator
to uint64, instead of bytes8. The problem with this is that it adds
semantic to MMInstantiator, making it responsible for understanding the
endianess of the values being passed. It makes more sense to leave it as
bytes8 and worry about the order of bytes on the emulator itself.
Changed MMInstantiator, MonotlithicRiscV and the monolithic test to
reflect this change.