Uncommented fetch_exception pma_start X flag
What was previsously thought to be a bug is not. It was a problem with
the dump log and not the step function. The X flag on the pma_start word
(the one that says if it is executable or not) was turned off so an
exception should indeed be raised. Uncommented the fetch_exception but
also turned the pma_start word from 4097 to 4129 (turned on the X flag).