Created read_x and write_x functions
Invoking mi.memoryRead directly with an address is error prone, making
the semantic confusing of wether youre writing/reading using the index
of the register or its address. read_x and write_x will translate from
the index to address, so rest of the code doesnt need to worry about it.